Fixed -Wenum-compare warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@91674 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/test/test_compress.cpp
@@ -383,11 +383,13 @@ void CTest::TestEmptyInputData(CCompressStream::EMethod method)
     size_t n;
 
     assert(
-        CZipCompression::fAllowEmptyData == CBZip2Compression::fAllowEmptyData
+        static_cast<int>(CZipCompression::fAllowEmptyData) == 
+        static_cast<int>(CBZip2Compression::fAllowEmptyData)
     );
 #ifdef HAVE_LIBLZO
     assert(
-        CZipCompression::fAllowEmptyData == CLZOCompression::fAllowEmptyData
+        static_cast<int>(CZipCompression::fAllowEmptyData) == 
+        static_cast<int>(CLZOCompression::fAllowEmptyData)
     );
 #endif
